Refactoring dialogs
