+++ app/src/main/java/com/example/habitsmasher/HabitEventDialog.java
@@ -1,12 +1,22 @@
 package com.example.habitsmasher;
 
+import android.app.DatePickerDialog;
+import android.content.Context;
+import android.net.Uri;
+import android.util.Log;
+import android.widget.Button;
+import android.widget.DatePicker;
 import android.widget.EditText;
+import android.widget.ImageView;
 import android.widget.TextView;
 
+import androidx.annotation.NonNull;
 import androidx.fragment.app.DialogFragment;
 
+import com.example.habitsmasher.ui.history.HabitEventListFragment;
+
 /**
- * HabitEventDialog is an abstract class that describes any dialog
+ * Abstract UI class that describes any dialog
  * involving adding and editing Habit Events
  * @author Jason Kim
  */
@@ -16,12 +26,77 @@ public abstract class HabitEventDialog extends DialogFragment implements Display
     public static final int INCORRECT_COMMENT = 1;
     public static final int INCORRECT_DATE = 2;
 
+    protected HabitEventListFragment _habitEventListFragment;
+    protected String TAG;
+    protected Uri _selectedImage;
+
+    // text elements in habit event dialog
     protected EditText _eventCommentText;
     protected TextView _eventDateText;
-
-    /* error text is a text view that has initially no content, and only gets a string value when
-      to display an error message */
     protected TextView _errorText;
+    protected TextView _header;
+
+    protected ImageView _eventPictureView;
+
+    protected Button _confirmButton;
+    protected Button _cancelButton;
+
+    /**
+     * Initializes the variables holding the UI elements
+     * of the habit event dialog
+     * @param view view representing the habit event dialog
+     */
+    protected void initializeUIElements(View view) {
+        _eventCommentText = view.findViewById(R.id.add_habit_event_comment);
+        _eventDateText = view.findViewById(R.id.habit_event_date_selection);
+        _errorText = view.findViewById(R.id.error_text_event);
+        _header = view.findViewById(R.id.add_habit_event_header);
+        _eventPictureView = view.findViewById(R.id.habit_event_add_photo);
+        _confirmButton = view.findViewById(R.id.confirm_habit_event);
+        _cancelButton = view.findViewById(R.id.cancel_habit_event);
+    }
+
+    protected void setDateTextViewListener() {
+        _eventDateText.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                openDatePickerDialog();
+            }
+        });
+    }
+
+    protected void setCancelButtonListener() {
+        _cancelButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                _errorText.setText("");
+                Log.d(TAG, "Cancel");
+                getDialog().dismiss();
+            }
+        });
+    }
+
+    /**
+     * Opens the calendar dialog used for date selection
+     */
+    protected void openDatePickerDialog() {
+        DatePickerDialogFragment datePickerDialogFragment = new DatePickerDialogFragment(new DatePickerDialog.OnDateSetListener() {
+            /**
+             * Sets the text of the date select view to reflect selected date
+             * @param view
+             * @param year year of selected date
+             * @param month month of selected date (integer from 0 to 11)
+             * @param day day of month of selected date
+             */
+            @Override
+            public void onDateSet(DatePicker view, int year, int month, int day) {
+                int correctedMonth = month + 1;
+                String date = day + "/" + correctedMonth + "/" + year;
+                _eventDateText.setText(date);
+            }
+        });
+        datePickerDialogFragment.show(getFragmentManager(), "DatePickerDialogFragment");
+    }
 
     @Override
     public void displayErrorMessage(int messageType) {
@@ -34,6 +109,18 @@ public abstract class HabitEventDialog extends DialogFragment implements Display
             case INCORRECT_DATE:
                 _errorText.setText("Please enter a start date");
                 break;
+        }
+    }
+
+    @Override
+    public void onAttach(@NonNull Context context) {
+        super.onAttach(context);
+        try {
+            // Get habit event fragment for later use
+            _habitEventListFragment = (HabitEventListFragment) getTargetFragment();
+        }
+        catch (ClassCastException e) {
+            Log.e(TAG, "Exception" + e.getMessage());
         }
     }
 }
